ci: enforce valid the pull request titles (#5)
# Description

Make sure that the title of the pull request matches the rules from
[Conventional
Commit](https://www.conventionalcommits.org/en/v1.0.0/#summary). This
makes it easy to create a release documentation and ensures a minimum
standard for the commit messages (together with the pull request
template).

The commit merged to the `main` branch is a squashed commit from the
feature branch. Title and description for this new commit are taken from
the pull request.

# Checklist

- [x] My code follows the style guidelines of the project
- [x] I have performed a self-review of my own code
- [ ] I have made corresponding changes to the documentation